move all math to math folder
